requirements.txt
pypdf
pytesseract
Pillow
